fix/#129 버튼 입력 시간 뷰모델에서 관리하도록
